+++ .circleci/config.yml
@@ -1420,7 +1420,7 @@ workflows:
             filters:
               branches:
                 ignore:
-                  - support-multitouch-gc-work
+                  - NONE
             workflow-name: "Continuous-integration"
         - sonar-check:
             context: SonarCloud
