+++ hedera-node/src/main/resources/bootstrap.properties
@@ -69,7 +69,7 @@ hedera.transaction.maxMemoUtf8Bytes=100
 hedera.transaction.maxValidDuration=180
 hedera.transaction.minValidDuration=15
 hedera.transaction.minValidityBufferSecs=10
-autorenew.isEnabled=true
+autorenew.isEnabled=false
 autoRenew.targetTypes=CONTRACT
 autorenew.numberOfEntitiesToScan=100
 autorenew.maxNumberOfEntitiesToRenewOrDelete=2
