+++ hedera-node/src/test/java/com/hedera/services/context/properties/BootstrapPropertiesTest.java
@@ -151,7 +151,7 @@ class BootstrapPropertiesTest {
 			entry("ledger.tokenTransfers.maxLen", 10),
 			entry("ledger.totalTinyBarFloat", 5000000000000000000L),
 			entry("autoCreation.enabled", true),
-			entry("autorenew.isEnabled", true),
+			entry("autorenew.isEnabled", false),
 			entry("autoRenew.targetTypes", EnumSet.of(CONTRACT)),
 			entry("autorenew.numberOfEntitiesToScan", 100),
 			entry("autorenew.maxNumberOfEntitiesToRenewOrDelete", 2),
